% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Broadstone_prey.R
\docType{data}
\name{Broadstone.prey}
\alias{Broadstone.prey}
\title{Macroinvertebrate abundance data to accompany the \code{Broadstone} data frame}
\format{A data frame with 19 columns and one row, with each column name
  representing one taxon. Names match those in \code{Broadstone}.}
\source{
Woodward, G., Speirs, D.C. & Hildrew, A.G. (2005) Quantification and
  resolution of a complex, size-structured food web. \emph{Advances in
  Ecological Research}, \strong{36}, 84--135.
}
\usage{
Broadstone.prey
}
\description{
The abundance of the 19 macroinvertebrate taxa present in the Broadstone food
  web in August 1996 (see Woodward \emph{et al}. 2005 for full details).
  The values are the total number of individuals collected from 30 Surber
  samples.
}
\keyword{datasets}

