% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silene_data_plants.R
\docType{data}
\name{Silene.plants}
\alias{Silene.plants}
\title{Flower abundance data to accompany the \emph{Silene} visitation network}
\format{
A data frame with 11 rows and six columns. Each row represents the
  floral abundance on one sampling occasion. The first column (\code{Visit})
  lists the codes for the sampling visit (V1--V11), whilst the remaining five
  variables represent the abundance of each of the flower species.
}
\source{
Gibson, R.H., Nelson, I.L., Hopkins, G.W., Hamlett, B.J. &
  Memmott, J. (2006) Pollinator webs, plant communities and the conservation
  of rare plants: arable weeds as a case study. \emph{Journal of Applied
  Ecology}, \strong{43}, 246--257.
}
\usage{
Silene.plants
}
\description{
Floral abundance data for the five flower species recorded in the Silene flower
  visitation network (see \code{\link{Silene}}) for full details.
}
\keyword{datasets}
