% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WelshStreams_prey.R
\docType{data}
\name{WelshStreams.prey}
\alias{WelshStreams.prey}
\title{Macroinvertebrate abundance to accompany part of the food web from upland
  streams in south Wales, UK}
\format{A data frame with 6 rows and 17 variables. Each row represents the
  abundance of individual macroinvertebrate taxa in each of the six streams.
  Column 1 contains the stream codes, whilst columns 2--17 represent the
  abundance of each macroinvertebrate taxon summed across three 1-minute
  kick-samples in each stream.}
\source{
Pearson, C.E., Symondson, W.O.C., Clare, E.L., Ormerod, S.J.,
 Iparraguirre Bolanos, E. & Vaughan, I.P. (2018) The effects of pastoral
 intensification on the feeding interactions of generalist predators in
 streams. \emph{Molecular Ecology}, \strong{27}, 590-602.
}
\usage{
WelshStreams.prey
}
\description{
The abundance of 16 potential macroinvertebrate prey taxa to accompany the
  food web data in \code{\link{WelshStreams}}.
}
\keyword{datasets}

